backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bid-agent-backend"
version = "0.1.0"
description = "Lightweight backend skeleton for the bid agent MVP."
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.116.0",
  "uvicorn[standard]>=0.35.0",
  "pydantic-settings>=2.10.0",
  "sqlalchemy>=2.0.38",
  "openai>=1.0.0",
  "psycopg[binary]>=3.2.0",
  "redis>=6.2.0",
  "python-multipart>=0.0.20",
  "pypdf>=5.4.0",
  "python-docx>=1.1.2",
  "minio>=7.2.7",
  "pymupdf>=1.26.0",
  "ocrmac>=1.0.0",
  "rapidocr_onnxruntime>=1.4.4",
]

[tool.setuptools.packages.find]
include = ["app*"]

[tool.ruff]
line-length = 100
target-version = "py311"
